+++ lifecycle/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.github.mag0716.arch.lifecyclesample">
 
     <application
+        android:name=".LoggingApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -12,10 +13,12 @@
         <activity android:name="com.github.mag0716.arch.lifecyclesample.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name=".SecondActivity" />
+
         <service android:name=".LoggingService" />
     </application>
