+++ lifecycle/src/main/java/com/github/mag0716/arch/lifecyclesample/MainActivity.java
@@ -1,18 +1,16 @@
 package com.github.mag0716.arch.lifecyclesample;
 
-import android.arch.lifecycle.Lifecycle;
 import android.arch.lifecycle.LifecycleObserver;
-import android.arch.lifecycle.LifecycleOwner;
-import android.arch.lifecycle.OnLifecycleEvent;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.View;
 
 import hugo.weaving.DebugLog;
 
 public class MainActivity extends AppCompatActivity {
 
-    private TestObserver lifecycleObserver = new TestObserver();
+    private LifecycleObserver lifecycleObserver = new LoggingObserver();
     private Intent serviceIntent;
 
     @DebugLog
@@ -43,41 +41,7 @@ public class MainActivity extends AppCompatActivity {
         getLifecycle().removeObserver(lifecycleObserver);
     }
 
-    static class TestObserver implements LifecycleObserver {
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_CREATE)
-        void onCreate(LifecycleOwner owner) {
-        }
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_START)
-        void onStart(LifecycleOwner owner) {
-        }
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
-        void onResume(LifecycleOwner owner) {
-        }
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_PAUSE)
-        void onPause(LifecycleOwner owner) {
-        }
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
-        void onStop(LifecycleOwner owner) {
-        }
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
-        void onDestroy(LifecycleOwner owner) {
-        }
-
-        @DebugLog
-        @OnLifecycleEvent(Lifecycle.Event.ON_ANY)
-        void onAny(LifecycleOwner owner, Lifecycle.Event event) {
-        }
+    public void onClick(View view) {
+        startActivity(new Intent(this, SecondActivity.class));
     }
 }
