+++ lifecycle/src/main/java/com/github/mag0716/arch/lifecyclesample/SecondActivity.java
@@ -1,0 +1,17 @@
+package com.github.mag0716.arch.lifecyclesample;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+
+/**
+ * Created by mag0716 on 2018/03/17.
+ */
+public class SecondActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_second);
+    }
+}
