+++ lifecycle/src/main/res/layout/activity_second.xml
@@ -8,13 +8,12 @@
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.github.mag0716.arch.lifecyclesample.MainActivity">
+        tools:context="com.github.mag0716.arch.lifecyclesample.SecondActivity">
 
-        <Button
+        <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:onClick="onClick"
-            android:text="Move to SecondActivity"
+            android:text="Hello World!"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
